refactor: Refactored observable folder structure
