+++ src/vidyano/index.ts
@@ -3,7 +3,7 @@ export * from "./common/bignumber.js"
 export * from "./common/boolean.js"
 export * from "./common/date.js"
 export * from "./common/expression-parser.js"
-export * from "./common/observable.js"
+export * from "./observable"
 export * from "./common/queue.js"
 export * from "./common/service-bus.js"
 export * from "./common/string.js"
